backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pgvector
sentence-transformers
google-generativeai
numpy
sqlalchemy
python-jose